#19 add CanPay/CanPayFloat
