+++ changelog.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- Add trait `CanPay`.
+- Add trait `CanPayFloat`.
+
+### Deprecated
+- Trait `CanBePaid`.
+- Trait `CanBePaidFloat`.
 
 ## [2.1.0] - 2018-11-22
 ### Added
